+++ Back-End-Maven/almoxarifado/src/main/java/br/senai/sc/almoxarifado/repository/LocalizacaoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LocalizacaoRepository extends JpaRepository<Localizacao, Integer> {
+public interface LocalizacaoRepository extends JpaRepository<Localizacao, Long> {
 }
